requirements.txt
torch
torchaudio
librosa
soundfile
numpy
pandas
scikit-learn
matplotlib
tqdm
pyyaml
utils